refactor: Overhaul project structure and dependencies
- Restructure project into a 'core' app for settings and a consolidated 'api' app.
- Remove unused Django apps and static files to clean up the project.
- Integrate Swagger for API documentation.
- Remove unused dependencies and regenerate requirements.txt.
