--- .envs
@@ -1,5 +1,0 @@
-DB_NAME= "verceldb"
-DB_USER= "default"
-DB_PASSWORD= "ONxhQ4tCbg7d"
-DB_HOST= "ep-falling-wildflower-a4im662j-pooler.us-east-1.aws.neon.tech"
-DB_PORT= 5432
