+++ api/models.py
@@ -1,5 +1,8 @@
 from django.db import models
-
+from django.conf import settings
+from django.db.models.signals import post_save
+from django.dispatch import receiver
+from rest_framework.authtoken.models import Token
 
 class Landlord(models.Model):
     landlord_name = models.CharField(max_length=30)
@@ -79,3 +82,9 @@ class MaintenanceRequest(models.Model):
 
     def __str__(self):
         return str(self.id)
+
+
+@receiver(post_save, sender=settings.AUTH_USER_MODEL)
+def create_auth_token(sender, instance=None, created=False, **kwargs):
+    if created:
+        Token.objects.create(user=instance)
