+++ api/user/views.py
@@ -2,8 +2,7 @@ from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from rest_framework.authtoken.models import Token
 from rest_framework import status
-from user_app.api.serializers import RegistrationSerializer
-from user_app import models
+from api.user.serializers import RegistrationSerializer
 
 @api_view(['POST',])
 def logout_view(request):
