--- render.yaml
@@ -1,22 +1,0 @@
-databases:
-  - name: plotidb
-    plan: free
-    databaseName: mysite
-    user: mysite
-
-services:
-  - type: web
-    plan: free
-    name: mysite
-    runtime: python
-    buildCommand: "./build.sh"
-    startCommand: "python -m gunicorn mysite.asgi:application -k uvicorn.workers.UvicornWorker"
-    envVars:
-      - key: DATABASE_URL
-        fromDatabase:
-          name: plotidb
-          property: connectionString
-      - key: SECRET_KEY
-        generateValue: true
-      - key: WEB_CONCURRENCY
-        value: 4
