--- vercel.json
@@ -1,13 +1,0 @@
-{
-    "builds": [{
-        "src": "mysite/wsgi.py",
-        "use": "@vercel/python",
-        "config": { "maxLambdaSize": "15mb", "runtime": "python3.9" }
-    }],
-    "routes": [
-        {
-            "src": "/(.*)",
-            "dest": "mysite/wsgi.py"
-        }
-    ]
-}
